Refactor: unified response, and messages
- Unified the response.
- Moved common messages to constants.py folder instead of system_messages that I created earlier, cause they serve the same purpose.
- updated the readme to include how to customize payload coming from clerk.
