--- app/utils/system_messages.py
@@ -1,1 +1,0 @@
-INVALID_BEARER_TOKEN_SCHEMA = "Missing or malformed Authorization header"
